Refactor DB using new Spring Data generated structure and integrate Hibernate annotations and runtime routine

+++ backend/src/main/java/com/codecool/dto/LocalTransactionDTO.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public record LocalTransactionDTO(
         String description,
         LocalDate dateOfTransaction,
-        int amount,
+        double amount,
         boolean isPlanned,
         boolean isRecurring
 ) {
